:hammer: refactor twitter button in component
