+++ frontend/src/components/home-buttons/types/social-media.tsx
@@ -1,0 +1,6 @@
+export interface ITwitter {
+  url: string
+  text?: string
+  via?: string
+  hashtags?: string
+}
