--- frontend/src/components/shortener-input/types/social-media.tsx
@@ -1,10 +1,0 @@
-interface ISocialMediaObjectKeys {
-  [key: string]: string | undefined;
-}
-
-export interface ISocialMedia extends ISocialMediaObjectKeys {
-  url: string
-  text?: string
-  via?: string
-  hashtags?: string
-}
